Separate depthcloud to object and viz
